+++ polaris-react/src/components/Button/Button.stories.tsx
@@ -8,7 +8,12 @@ import {
   VerticalStack,
   Text,
 } from '@shopify/polaris';
-import {PlusMinor, DeleteMinor} from '@shopify/polaris-icons';
+import {
+  PlusMinor,
+  DeleteMinor,
+  CancelSmallMinor,
+  EditMajor,
+} from '@shopify/polaris-icons';
 
 export default {
   component: Button,
@@ -28,6 +33,27 @@ export function All() {
               Label
             </Button>
             <Button disclosure>Label</Button>
+            <Button
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              icon={EditMajor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              disabled
+              icon={PlusMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              icon={DeleteMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
           <Card>
             <HorizontalStack gap="5" blockAlign="end">
@@ -38,6 +64,17 @@ export function All() {
                 Label
               </Button>
               <Button disclosure>Label</Button>
+              <Button
+                icon={CancelSmallMinor}
+                onClick={() => {}}
+                accessibilityLabel="Dismiss"
+              />
+              <Button
+                disabled
+                icon={EditMajor}
+                onClick={() => {}}
+                accessibilityLabel="Dismiss"
+              />
             </HorizontalStack>
           </Card>
         </VerticalStack>
@@ -57,6 +94,19 @@ export function All() {
             <Button destructive disclosure>
               Label
             </Button>
+            <Button
+              destructive
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              disabled
+              destructive
+              icon={EditMajor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
           <Card>
             <HorizontalStack gap="5" blockAlign="end">
@@ -73,6 +123,19 @@ export function All() {
               <Button destructive disclosure>
                 Label
               </Button>
+              <Button
+                destructive
+                icon={CancelSmallMinor}
+                onClick={() => {}}
+                accessibilityLabel="Dismiss"
+              />
+              <Button
+                disabled
+                destructive
+                icon={EditMajor}
+                onClick={() => {}}
+                accessibilityLabel="Dismiss"
+              />
             </HorizontalStack>
           </Card>
         </VerticalStack>
@@ -92,6 +155,19 @@ export function All() {
             <Button primary disclosure>
               Label
             </Button>
+            <Button
+              primary
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              disabled
+              primary
+              icon={EditMajor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
@@ -113,6 +189,13 @@ export function All() {
             <Button primary destructive disclosure>
               Label
             </Button>
+            <Button
+              primary
+              destructive
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
@@ -132,13 +215,18 @@ export function All() {
             <Button primarySuccess disclosure>
               Label
             </Button>
+            <Button
+              primarySuccess
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
         <VerticalStack gap="2">
           <Text as="h2">plain</Text>
           <HorizontalStack gap="5" blockAlign="end">
-            <Button plain>Label</Button>
             <Button plain>Label</Button>
             <Button plain disabled>
               Label
@@ -152,6 +240,22 @@ export function All() {
             <Button plain disclosure>
               Label
             </Button>
+            <Button plain disclosure disabled>
+              Label
+            </Button>
+            <Button
+              plain
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              plain
+              disabled
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
@@ -173,6 +277,24 @@ export function All() {
             <Button plain destructive disclosure>
               Label
             </Button>
+            <Button plain destructive disclosure disabled>
+              Label
+            </Button>
+            <Button
+              plain
+              destructive
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              plain
+              destructive
+              disabled
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
@@ -194,6 +316,21 @@ export function All() {
             <Button primary plain disclosure>
               Label
             </Button>
+            <Button
+              primary
+              plain
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
+            <Button
+              disabled
+              primary
+              plain
+              icon={EditMajor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
 
@@ -215,6 +352,14 @@ export function All() {
             <Button primary plain destructive disclosure>
               Label
             </Button>
+            <Button
+              primary
+              plain
+              destructive
+              icon={CancelSmallMinor}
+              onClick={() => {}}
+              accessibilityLabel="Dismiss"
+            />
           </HorizontalStack>
         </VerticalStack>
       </VerticalStack>
@@ -292,7 +437,22 @@ export function Slim() {
 }
 
 export function Large() {
-  return <Button size="large">Create store</Button>;
+  return (
+    <VerticalStack gap="4">
+      <HorizontalStack gap="5" blockAlign="end">
+        <Button size="large">Create store</Button>
+        <Button size="large" accessibilityLabel="Edit" icon={EditMajor} />
+        <Button size="large" icon={PlusMinor} primary>
+          Create store
+        </Button>
+      </HorizontalStack>
+      <HorizontalStack gap="5" blockAlign="end">
+        <Button size="large" fullWidth>
+          Create store
+        </Button>
+      </HorizontalStack>
+    </VerticalStack>
+  );
 }
 
 export function FullWidth() {