+++ polaris-react/src/components/TopBar/components/UserMenu/UserMenu.tsx
@@ -7,6 +7,7 @@ import {MessageIndicator} from '../../../MessageIndicator';
 import {Menu} from '../Menu';
 import type {MenuProps} from '../Menu';
 import {Text} from '../../../Text';
+import {useFeatures} from '../../../../utilities/features';
 
 import styles from './UserMenu.scss';
 
@@ -49,6 +50,7 @@ export function UserMenu({
   customWidth,
 }: UserMenuProps) {
   const showIndicator = Boolean(message);
+  const {polarisSummerEditions2023} = useFeatures();
 
   const activatorContentMarkup = customActivator ? (
     customActivator
@@ -58,20 +60,22 @@ export function UserMenu({
         <Text as="p" alignment="start" fontWeight="medium" truncate>
           {name}
         </Text>
-        <Text
-          as="p"
-          variant="bodySm"
-          alignment="start"
-          color="subdued"
-          truncate
-        >
-          {detail}
-        </Text>
+        <span className={styles.Message}>
+          <Text
+            as="p"
+            variant="bodySm"
+            alignment="start"
+            color="subdued"
+            truncate
+          >
+            {detail}
+          </Text>
+        </span>
       </span>
       <MessageIndicator active={showIndicator}>
         <Avatar
           shape="square"
-          size="small"
+          size={polarisSummerEditions2023 ? 'extraSmall' : 'small'}
           initials={initials && initials.replace(' ', '')}
           source={avatar}
         />
@@ -89,6 +93,7 @@ export function UserMenu({
       message={message}
       accessibilityLabel={accessibilityLabel}
       customWidth={customWidth}
+      userMenu
     />
   );
 }